Move `BaseItemList` to `Icinga\Module\Icingadb\Common`
