+++ library/Icingadb/Web/Controller.php
@@ -10,7 +10,7 @@ use Icinga\Application\Icinga;
 use Icinga\Module\Icingadb\Common\Auth;
 use Icinga\Module\Icingadb\Common\Database;
 use Icinga\Module\Icingadb\Web\Control\SearchBar\ObjectSuggestions;
-use Icinga\Module\Icingadb\Widget\BaseItemList;
+use Icinga\Module\Icingadb\Common\BaseItemList;
 use Icinga\Module\Icingadb\Widget\ViewModeSwitcher;
 use Icinga\Security\SecurityException;
 use InvalidArgumentException;
